+++ src/polyfills/setCustomValidity.js
@@ -2,6 +2,8 @@
 
 
 import { message_store } from '../components/message_store';
+import Renderer from '../components/renderer';
+import ValidityState from './validityState';
 
 
 /**
@@ -9,4 +11,11 @@ import { message_store } from '../components/message_store';
  */
 export default function setCustomValidity(element, msg) {
   message_store.set(element, msg, true);
+  /* live-update the warning */
+  const warning = Renderer.getWarning(element);
+  if (warning) {
+    Renderer.setMessage(warning, msg, element);
+  }
+  /* update any classes if the validity state changes */
+  ValidityState(element).valid;
 }